refactor: HouseWork 관련 엔티티, 리포지토리 이동
