+++ src/main/java/team1/allo/housework/repository/housework/HouseWorkCustomRepository.java
@@ -1,9 +1,9 @@
-package team1.allo.group.repository.housework;
+package team1.allo.housework.repository.housework;
 
 import java.time.LocalDate;
 import java.util.List;
 
-import team1.allo.group.entity.HouseWork;
+import team1.allo.housework.entity.HouseWork;
 
 public interface HouseWorkCustomRepository {
