+++ src/main/java/team1/allo/housework/repository/housework/HouseWorkRepository.java
@@ -1,11 +1,11 @@
-package team1.allo.group.repository.housework;
+package team1.allo.housework.repository.housework;
 
 import java.time.LocalDate;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import team1.allo.group.entity.HouseWork;
+import team1.allo.housework.entity.HouseWork;
 
 public interface HouseWorkRepository extends JpaRepository<HouseWork, Long>, HouseWorkCustomRepository {
