+++ src/main/java/team1/allo/housework/repository/houseworkmember/HouseWorkMemberCustomRepositoryImpl.java
@@ -1,4 +1,4 @@
-package team1.allo.group.repository.houseworkmember;
+package team1.allo.housework.repository.houseworkmember;
 
 import static team1.allo.group.entity.QHouseWorkMember.*;
 
@@ -7,7 +7,7 @@ import java.util.List;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import lombok.RequiredArgsConstructor;
-import team1.allo.group.entity.HouseWorkMember;
+import team1.allo.housework.entity.HouseWorkMember;
 
 @RequiredArgsConstructor
 public class HouseWorkMemberCustomRepositoryImpl implements HouseWorkMemberCustomRepository {
