+++ src/main/java/team1/allo/housework/repository/houseworkmember/HouseWorkMemberRepository.java
@@ -1,8 +1,8 @@
-package team1.allo.group.repository.houseworkmember;
+package team1.allo.housework.repository.houseworkmember;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import team1.allo.group.entity.HouseWorkMember;
+import team1.allo.housework.entity.HouseWorkMember;
 
 public interface HouseWorkMemberRepository
 	extends JpaRepository<HouseWorkMember, Long>, HouseWorkMemberCustomRepository {
