+++ src/main/java/team1/allo/housework/repository/houseworktag/HouseWorkTagRepository.java
@@ -1,8 +1,8 @@
-package team1.allo.group.repository.houseworktag;
+package team1.allo.housework.repository.houseworktag;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import team1.allo.group.entity.HouseWorkTag;
+import team1.allo.housework.entity.HouseWorkTag;
 
 public interface HouseWorkTagRepository extends JpaRepository<HouseWorkTag, Long>, HouseWorkTagCustomRepository {
 }
